+++ src/io/dss_parse.jl
@@ -1032,7 +1032,7 @@ function _apply_like!(raw_dss, dss_data, comp_type)
             end
 
             if prop in links
-                linked_dss = _find_component(dss_data, raw_dss[prop], comp_type)
+                linked_dss = find_component(dss_data, raw_dss[prop], comp_type)
                 if isempty(linked_dss)
                     Memento.warn(_LOGGER, "$comp_type.$(raw_dss["name"]): $prop=$(raw_dss[prop]) cannot be found")
                 else