+++ test/data.jl
@@ -57,4 +57,14 @@
             @test isapprox(A, B)
         end
     end
+
+    @testset "node counting functions" begin
+        dss = PMD.parse_dss("../test/data/opendss/case5_phase_drop.dss")
+        pmd = PMD.parse_file("../test/data/opendss/case5_phase_drop.dss")
+        matlab = PMD.parse_file("../test/data/matlab/case5_i_r_b.m")
+
+        @test count_nodes(dss) == 7
+        @test count_nodes(dss) == count_nodes(pmd)
+        @test count_nodes(matlab) == 15
+    end
 end